refactor(user): :construction: make routes for user, fix service
Added the four user routes that are present in the javascript version, and changed method signatures and styling a bit to make them more readable. Also fixed a minor bug on the update method where the function was searching for the ID at the wrong place. Further filtered the documents being returned from the database to hide critical or unnecessary information.