refactor(models): :recycle: change constants to camelCase names, add new model, make interface name more descriptive
The community standard for JavaScript/TypeScript is to use camelCase for all variables and constants. Interfaces named <something>Optional were renamed to <something>Parameters, so as to make it clearer that it is not the object itself that is optional, but instead their parameters. Added 'organization.model' that is going to substitute the Ej model once the transition to TypeScript is complete.