fix: :bug: fix user object noise
Mongoose objects have a lot of hidden properties. If one does not use the 'lean( )' function, they are not converted to pure javascript, and thus will break if the application ever executes an operation on the object itself, as it is being done here to remove the password before sending the object to the frontend.